+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Automatically generate Tasks for the dashboard to remind editors to publish pages once translations land - @stevejalim
+  - BREAKING CHANGE: requires `django.contrib.humanize` to be added to `settings.INSTALLED_APPS`
 
 ## [0.9.0] - 2024-12-12